pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["elokow <lukas.kowejsza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
openai = "^1.12.0"
llama_index = "^0.10.6"
langchain = "^0.1.7"
jupyter = "^1.0.0"
chromadb = "0.4.22"
python-dotenv = "1.0.1"
pydantic = "2.6.1"
pytest = "8.0.1"
langchain_openai = "0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
